Removed duplicate phonemes from the output
In version 0.8.0 all elements of the phonemes list were directly added to phones list and set was finally applied to remove duplicates. However phonemes list elements can contain several phonemes separated by a white space by default. Consequently if [a e] and [e a] were in phonemes then they were appended to phones and treated as different elements by "set" function. So the duplicates were NOT remove (in that case lower_phones would show "a e e a" instead of "a e").

That's why each phonemes element ph is split (white space is the default separator) and then the split result is appended to phones list.